% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_buffer-points-linear.R
\name{buffer_points_linear}
\alias{buffer_points_linear}
\title{Buffer points linear}
\usage{
buffer_points_linear(points, radius = 2, res = 8)
}
\arguments{
\item{points}{matrix, data frame
or object of class \code{sf} with \code{POINT} geometry}

\item{radius}{number of rings (used as buffer around points)}

\item{res}{resolution of hexagons to return}
}
\description{
Weight is calculated as \code{1 - distance / (1 + radius)}.
}
\examples{
library(sf)
library(leaflet)

bart_stations <- system.file("geojson/bart_stations.geojson", package = "h3forr") \%>\%
  st_read(quiet = TRUE)

buffer <- buffer_points_linear(bart_stations, radius = 2, res = 8)
hexagons <- h3_to_geo_boundary(buffer$h3_index) \%>\% geo_boundary_to_sf()

pal <- colorNumeric("Blues", buffer$norm)

map <- leaflet() \%>\% addTiles() \%>\%
  addPolygons(
    data = hexagons,
    fillOpacity = 0.9,
    fillColor = pal(buffer$norm),
    weight = 1,
    color = "white",
    label = sprintf("\%s", buffer$norm)
  ) \%>\%
  addMarkers(data = bart_stations)

if (interactive()) map
}
