% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_buffer-points-linear.R
\name{count_points}
\alias{count_points}
\title{Count points in each hexagon for a given resolution}
\usage{
count_points(points, res = 7)
}
\arguments{
\item{points}{matrix, data frame
or object of class \code{sf} with \code{POINT} geometry}

\item{res}{resolution of hexagons to return}
}
\description{
Count points in each hexagon for a given resolution
}
