% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_info.R
\name{edge_length}
\alias{edge_length}
\title{Get the average hexagon edge length at a given resolution}
\usage{
edge_length(res = 7, unit = "m")
}
\arguments{
\item{res}{resolution between 0 and 15 [numeric vector]}

\item{unit}{unit, either \code{m} for meters or \code{km} for kilometers}
}
\value{
numeric vector
}
\description{
js pendant: \code{h3.edgeLength(res, unit) => number}
}
\examples{
edge_length(1:4, unit = "km")

}
