% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-boundary-to-sf.R
\name{geo_boundary_to_sf}
\alias{geo_boundary_to_sf}
\alias{geo_boundary_to_sf.matrix}
\alias{geo_boundary_to_sf.list}
\title{Parse geo boundaries to object of class \code{sf}}
\usage{
geo_boundary_to_sf(obj)

\method{geo_boundary_to_sf}{matrix}(obj)

\method{geo_boundary_to_sf}{list}(obj)
}
\arguments{
\item{obj}{geo boundary matrix or list of matrices with [lng, lat] pairs
as returned from \link{h3_to_geo_boundary} or \link{h3_set_to_multi_polygon}}
}
\value{
object of class \code{sf} (geometry type: \code{POLYGON})
}
\description{
Parse geo boundaries to object of class \code{sf}
}
