% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_geo-to-h3.R
\name{geo_to_h3}
\alias{geo_to_h3}
\title{Get the hexagon containing a lat,lon point}
\usage{
geo_to_h3(lat, lng, res)
}
\arguments{
\item{lat}{latidude}

\item{lng}{longitude}

\item{res}{resolution of hexagons to return}
}
\description{
Get the hexagon containing a lat,lon point
}
