% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-to-sf.R
\name{geo_to_sf}
\alias{geo_to_sf}
\alias{geo_to_sf.matrix}
\alias{geo_to_sf.numeric}
\alias{geo_to_sf.list}
\title{Parse [lat, lng] points to sf object}
\usage{
geo_to_sf(coords)

\method{geo_to_sf}{matrix}(obj)

\method{geo_to_sf}{numeric}(coords)

\method{geo_to_sf}{list}(coords)
}
\arguments{
\item{coords}{[lat, lng] points}
}
\description{
Parse [lat, lng] points to sf object
}
