% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_edges.R
\name{get_destination_h3_index_from_unidirectional_edge}
\alias{get_destination_h3_index_from_unidirectional_edge}
\title{Get the destination H3 (hexagon) index from an H3 edge index}
\usage{
get_destination_h3_index_from_unidirectional_edge(h3_edge_index)
}
\arguments{
\item{h3_edge_index}{character vector representing H3 edge indexes}
}
\value{
character vector
}
\description{
Get the destination H3 (hexagon) index from an H3 edge index
}
