% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_edges.R
\name{get_h3_indexes_from_unidirectional_edge}
\alias{get_h3_indexes_from_unidirectional_edge}
\title{Get the H3 [origin, destination] index pair from an H3 edge index}
\usage{
get_h3_indexes_from_unidirectional_edge(h3_edge_index)
}
\arguments{
\item{h3_edge_index}{H3 edge index [vector]}
}
\description{
Get the H3 [origin, destination] index pair from an H3 edge index
}
