% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_edges.R
\name{get_h3_unidirectional_edge}
\alias{get_h3_unidirectional_edge}
\title{Get an H3 unidirectional edge index}
\usage{
get_h3_unidirectional_edge(origin_h3_index, destination_h3_index)
}
\arguments{
\item{origin_h3_index}{H3 index (origin) [vector of indexes]}

\item{destination_h3_index}{H3 index (destination) [single index]}
}
\value{
H3 edge index vector
}
\description{
Get an H3 edge index for a given origin and destination H3 index.
}
\note{
js pendant: \code{h3.getH3UnidirectionalEdge(origin, destination) => H3 edge index}
}
