% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_edges.R
\name{get_h3_unidirectional_edge_boundary}
\alias{get_h3_unidirectional_edge_boundary}
\title{Get the vertices of a given edge as an array of [lat, lng] points.}
\usage{
get_h3_unidirectional_edge_boundary(h3_edge_index,
  format_as_geojson = TRUE)
}
\arguments{
\item{h3_edge_index}{H3 edge index [character vector]}

\item{format_as_geojson}{provide GeoJSON output: [lng, lat] and closed loops?}
}
\description{
Get the vertices of a given edge as an array of [lat, lng] points.
}
\note{
Edges that cross the edge of an icosahedron face may return 3 coordinates.
}
