% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_edges.R
\name{get_h3_unidirectional_edges_from_hexagon}
\alias{get_h3_unidirectional_edges_from_hexagon}
\title{Get all unidirectional edges for a given H3 index (an edge to every neighbor)}
\usage{
get_h3_unidirectional_edges_from_hexagon(h3_index)
}
\arguments{
\item{h3_index}{H3 index [vector]}
}
\value{
H3 edge index vector or list of H3 edge index vectors
}
\description{
js pendant: \code{h3.getH3UnidirectionalEdgesFromHexagon(h3Index) => Array.<H3EdgeIndex>}
}
\examples{
get_sample_h3_index() \%>\%
  get_h3_unidirectional_edges_from_hexagon() \%>\%
  h3_unidirectional_edge_is_valid()

}
