% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_edges.R
\name{get_origin_h3_index_from_unidirectional_edge}
\alias{get_origin_h3_index_from_unidirectional_edge}
\title{Get the origin H3 (hexagon) index from an H3 edge index}
\usage{
get_origin_h3_index_from_unidirectional_edge(h3_edge_index)
}
\arguments{
\item{h3_edge_index}{H3 edge index [vector]}
}
\value{
H3 index vector
}
\description{
js pendant: \code{h3.getOriginH3IndexFromUnidirectionalEdge(edgeIndex) => H3 index}
}
