% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{greater_manchester}
\alias{greater_manchester}
\title{greater manchester}
\format{a matrix with 216 rows and 2 columns:
\describe{
  \item{lng}{longitude}
  \item{lat}{latitude}
}}
\source{
\url{http://download.geofabrik.de/europe/great-britain/england/greater-manchester.poly}
}
\usage{
greater_manchester
}
\description{
Coordinates describing the polygon of greater manchester
}
\keyword{datasets}
