% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_core.R
\name{h3_get_base_cell}
\alias{h3_get_base_cell}
\title{Get the number of the base cell for a given H3 index}
\usage{
h3_get_base_cell(h3_index)
}
\arguments{
\item{h3_index}{character vector representing H3 indexes}
}
\value{
numeric vector representing indexes of the base cells (1 - 122)
}
\description{
Get the number of the base cell for a given H3 index
}
