% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_core.R
\name{h3_get_resolution}
\alias{h3_get_resolution}
\title{Get the resolution of an H3 index}
\usage{
h3_get_resolution(h3_index)
}
\arguments{
\item{h3_index}{H3 index [vector]}
}
\value{
numeric vector
}
\description{
Get the resolution of an H3 index
}
