% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_core.R
\name{h3_is_pentagon}
\alias{h3_is_pentagon}
\title{Check whether the given H3 index is a pentagon}
\usage{
h3_is_pentagon(h3_index)
}
\arguments{
\item{h3_index}{character vector representing H3 indexes}
}
\value{
boolean vector
}
\description{
Check whether the given H3 index is a pentagon
}
