% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_core.R
\name{h3_is_res_class_iii}
\alias{h3_is_res_class_iii}
\title{Check whether the given H3 index is in a Class III resolution}
\usage{
h3_is_res_class_iii(h3_index)
}
\arguments{
\item{h3_index}{character vector representing H3 indexes}
}
\value{
boolean vector
}
\description{
Check whether the given H3 index is in a Class III resolution
(rotated versus the icosahedron and subject to shape distortion adding extra points
on icosahedron edges, making them not true hexagons).
}
