% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_algorithms2.R
\name{h3_to_children}
\alias{h3_to_children}
\title{Get the children/descendents of the given hexagon at a particular resolution}
\usage{
h3_to_children(h3_index, res = NULL)
}
\arguments{
\item{h3_index}{H3 index [vector]}

\item{res}{resolution (if \code{NULL} the resolution will be set to
\code{h3_get_resolution(h3_index[1]) + 1}}
}
\description{
Get the children/descendents of the given hexagon at a particular resolution
}
\examples{
coords <- c(37.36156, -122.05532)

h3_index <- geo_to_h3(coords)

h3_to_children(h3_index) \%>\%
  h3_to_geo_boundary() \%>\%
  geo_boundary_to_sf() \%>\% plot()

h3_to_geo_boundary(h3_index) \%>\%
  geo_boundary_to_sf() \%>\% plot(add = TRUE)
}
