% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_h3-to-geo.R
\name{h3_to_geo}
\alias{h3_to_geo}
\title{Get the [lat, lon] center of a given hexagon}
\usage{
h3_to_geo(h3_index)
}
\arguments{
\item{h3_index}{character vector representing H3 indexes}
}
\value{
numeric vector: [lat, lng];
matrix with [lat, lng] pairs if multiple indexes are supplied
}
\description{
Get the [lat, lon] center of a given hexagon
}
\examples{
quakes[1:3, c("lat", "long")] \%>\%
  geo_to_h3() \%>\%
  h3_to_geo()
}
