% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_hexbins.R
\name{hexbins}
\alias{hexbins}
\alias{hexbins.sf}
\alias{hexbins.data.frame}
\alias{hexbins.matrix}
\title{Get hexbins for given points and resolution}
\usage{
hexbins(data, res = 7, ...)

\method{hexbins}{sf}(data, res = 7, ...)

\method{hexbins}{data.frame}(data, res = 7, ..., lat = "lat",
  lng = "lng")

\method{hexbins}{matrix}(data, res = 7, ..., lat = "lat",
  lng = "lng")
}
\arguments{
\item{data}{matrix, data frame (with latidude and longitude columns)
or object of class \code{sf} (\code{POINT} geometry)}

\item{res}{resolution of hexagons to return; number between 0 and 15}

\item{...}{ignored}

\item{lat}{name of latitude column}

\item{lng}{name of longitude column}
}
\description{
Get hexbins for given points and resolution
}
