% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_utils.R
\name{hexbins_legacy}
\alias{hexbins_legacy}
\title{Get hexbins}
\usage{
hexbins_legacy(data, res = 7, lat = "lat", lng = "lng")
}
\arguments{
\item{data}{matrix, data frame or \code{sf} object}

\item{res}{resolution of hexagons to return; number between 0 and 15}

\item{lat}{name of latitude column if matrix or data frame is supplied}

\item{lng}{name of longitude column if matrix or data frame is supplied}
}
\description{
Get hexbins
}
