% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_k-ring.R
\name{k_ring}
\alias{k_ring}
\title{Get all hexagons in a k-ring around a given center}
\usage{
k_ring(h3_index, ring_size = 1)
}
\arguments{
\item{h3_index}{character vector representing H3 indexes}

\item{ring_size}{number of rings}
}
\value{
H3 indexes; list of character vectors if multiple centers are supplied
}
\description{
Get all hexagons in a k-ring around a given center
}
\note{
The order of the hexagons is undefined.
}
\examples{
coords <- c(37.3615593, -122.0553238)

geo_to_h3(coords) \%>\%
  k_ring() \%>\%
  h3_to_geo_boundary() \%>\%
  geo_boundary_to_sf() \%>\%
  plot()
}
