% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_algorithms.R
\name{k_ring_distances}
\alias{k_ring_distances}
\title{Get all hexagons in a k-ring around a given center, in an array of arrays ordered by distance from the origin.}
\usage{
k_ring_distances(h3_index, ring_size = 1)
}
\arguments{
\item{h3_index}{H3 indexes of center hexagons}

\item{ring_size}{number of rings}
}
\description{
The order of the hexagons within each ring is undefined.
}
\examples{
coords <- c(37.3615593, -122.0553238)
h3_index <- geo_to_h3(coords)

k_ring_distances(h3_index, 2)[[3]] \%>\%
  h3_to_geo_boundary() \%>\%
  geo_boundary_to_sf() \%>\%
  plot()
}
