% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manhattan}
\alias{manhattan}
\title{Persons in manhattan}
\format{a tibble with 13987 rows and 3 variables:
\describe{
  \item{lng}{longitude}
  \item{lat}{latitude}
  \item{sex}{1: male, 2: female}
}}
\source{
\url{https://raw.githubusercontent.com/uber-common/deck.gl-data/master/examples/scatterplot/manhattan.json}
}
\usage{
manhattan
}
\description{
Persons in manhattan
}
\keyword{datasets}
