% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_algorithms.R
\name{polyfill}
\alias{polyfill}
\title{Get all hexagons with centers contained in a given polygon.}
\usage{
polyfill(coords, res = 7, is_geojson = TRUE)
}
\arguments{
\item{coords}{[lat, lng] point}

\item{res}{resolution of hexagon to return}

\item{is_geojson}{expect GeoJson-style [lng, lat] pairs instead of [lat, lng]?}
}
\description{
The polygon is specified with GeoJson semantics as an array of loops.
Each loop is an array of [lat, lng] pairs (or [lng, lat] if isGeoJson is specified).
The first loop is the perimeter of the polygon, and subsequent loops are expected to be holes.
}
\examples{
library(sf)

nc <- system.file("shape/nc.shp", package = "sf") \%>\%
  st_read(quiet = TRUE)

st_geometry(nc[1, ]) \%>\% plot()
coords <- st_coordinates(nc[1, ])[, c("X", "Y")]

polyfill(coords) \%>\% h3_to_geo_boundary() \%>\%
  geo_boundary_to_sf() \%>\%
  plot(add = TRUE)
}
