% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_un-compact.R
\name{uncompact}
\alias{uncompact}
\title{Uncompact a compacted set of hexagons to hexagons of the same resolution}
\usage{
uncompact(h3_index, res = 7)
}
\arguments{
\item{h3_index}{character vector representing H3 indexes}

\item{res}{resolution to uncompact to; number between 0 and 15}
}
\value{
character vector of H3 indexes;
throws an error if an invalid resolution is supplied
}
\description{
Uncompact a compacted set of hexagons to hexagons of the same resolution
}
\seealso{
\code{\link{compact}}
}
