% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf-stuff.R
\name{geo_boundary_to_sf}
\alias{geo_boundary_to_sf}
\alias{geo_boundary_to_sf.default}
\alias{geo_boundary_to_sf.list}
\title{Parse geo boundary array(s) to sf object(s)}
\usage{
geo_boundary_to_sf(obj)

\method{geo_boundary_to_sf}{default}(obj)

\method{geo_boundary_to_sf}{list}(obj)
}
\arguments{
\item{obj}{geo boundary array(s)}
}
\description{
Parse geo boundary array(s) to sf object(s)
}
