% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_geo-to-h3.R
\name{geo_to_h3}
\alias{geo_to_h3}
\alias{geo_to_h3.default}
\alias{geo_to_h3.data.frame}
\alias{geo_to_h3.sf}
\title{Get the hexagons containing a [lat, lon] point}
\usage{
geo_to_h3(coords, res = 7)

\method{geo_to_h3}{default}(coords, res = 7)

\method{geo_to_h3}{data.frame}(coords, res = 7)

\method{geo_to_h3}{sf}(coords, res = 7)
}
\arguments{
\item{coords}{[lat, lng] point [vector, data.frame or sf object]}

\item{res}{resolution of hexagon to return}
}
\value{
H3 index vector
}
\description{
Get the hexagons containing a [lat, lon] point
}
\examples{
coords <- c(37.3615593, -122.0553238)
geo_to_h3(coords, 7)

geo_to_h3(quakes[1:3, c("lat", "long")], res = 5)
}
