% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_geo-to-h3.R
\name{geo_to_h3}
\alias{geo_to_h3}
\alias{geo_to_h3.numeric}
\alias{geo_to_h3.matrix}
\alias{geo_to_h3.data.frame}
\alias{geo_to_h3.sf}
\title{Get all hexagons containing [lat, lng] points}
\usage{
geo_to_h3(coords, res = 7)

\method{geo_to_h3}{numeric}(coords, res = 7)

\method{geo_to_h3}{matrix}(coords, res = 7)

\method{geo_to_h3}{data.frame}(coords, res = 7)

\method{geo_to_h3}{sf}(coords, res = 7)
}
\arguments{
\item{coords}{numeric vector, matrix or data.frame with [lat, lng] pairs;
object of class \code{sf} (geometry type: \code{POINT})}

\item{res}{resolution of hexagons to return; number between 0 and 15}
}
\value{
H3 indexes; character vector
}
\description{
Get all hexagons containing [lat, lng] points
}
\examples{
# numeric
coords <- c(37.3615593, -122.0553238)
geo_to_h3(coords, 7)

# matrix
geo_to_h3(road_safety_greater_manchester[1:10, ], res = 5)

# sf
points <- as.data.frame(greater_manchester) \%>\%
  sf::st_as_sf(coords = c("lng", "lat"), crs = 4326)
geo_to_h3(points[1:5, ], res = 12)
}
