% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-to-sf.R
\name{geo_to_sf}
\alias{geo_to_sf}
\alias{geo_to_sf.matrix}
\alias{geo_to_sf.numeric}
\alias{geo_to_sf.list}
\title{Parse [lat, lng] points to object of class \code{sf}}
\usage{
geo_to_sf(lat_lng_obj)

\method{geo_to_sf}{matrix}(lat_lng_obj)

\method{geo_to_sf}{numeric}(lat_lng_obj)

\method{geo_to_sf}{list}(lat_lng_obj)
}
\arguments{
\item{lat_lng_obj}{numeric vector or matrix with [lat, lng] points as returned
from \link{h3_to_geo} or \link{h3_to_geo_boundary} with \code{format_as_geojson = FALSE};
a list of matrices is also supported}
}
\value{
object of class \code{sf} (geometry type: \code{POINT})
}
\description{
Parse [lat, lng] points to object of class \code{sf}
}
\examples{
coords <- c(37.3615593, -122.0553238)

h3_index <- geo_to_h3(coords, 7)
h3_indexes <- k_ring(h3_index)

# Single center point
h3_to_geo(h3_index) \%>\% geo_to_sf()

# Multiple center points
h3_to_geo(h3_indexes) \%>\% geo_to_sf()

# Boundary points (single hexagon)
h3_to_geo_boundary(h3_index, format_as_geojson = FALSE) \%>\%
  geo_to_sf()

# Boundary points (multiple hexagons)
h3_to_geo_boundary(h3_indexes, format_as_geojson = FALSE) \%>\%
  geo_to_sf()
}
