% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_edges.R
\name{get_h3_unidirectional_edge}
\alias{get_h3_unidirectional_edge}
\title{Get an H3 edge index for a given origin and destination H3 (hexagon) index}
\usage{
get_h3_unidirectional_edge(origin_h3_index, destination_h3_index)
}
\arguments{
\item{origin_h3_index}{H3 index (origin); character vector}

\item{destination_h3_index}{H3 index (destination); character (single value)}
}
\value{
character vector
}
\description{
Get an H3 edge index for a given origin and destination H3 (hexagon) index
}
