% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_edges.R
\name{get_h3_unidirectional_edges_from_hexagon}
\alias{get_h3_unidirectional_edges_from_hexagon}
\title{Get all unidirectional edges for a given H3 index (an edge to every neighbor)}
\usage{
get_h3_unidirectional_edges_from_hexagon(h3_index)
}
\arguments{
\item{h3_index}{character vector representing H3 indexes}
}
\value{
chraracter vector;
list of character vectors if multiple indexes are supplied
}
\description{
Get all unidirectional edges for a given H3 index (an edge to every neighbor)
}
\examples{
get_sample_h3_index() \%>\%
  get_h3_unidirectional_edges_from_hexagon() \%>\%
  h3_unidirectional_edge_is_valid()

}
