% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_core.R
\name{h3_get_resolution}
\alias{h3_get_resolution}
\title{Get the resolution of a given H3 index}
\usage{
h3_get_resolution(h3_index)
}
\arguments{
\item{h3_index}{character vector representing H3 indexes}
}
\value{
numeric vector with resolutions
}
\description{
Get the resolution of a given H3 index
}
