% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_edges.R
\name{h3_indexes_are_neighbors}
\alias{h3_indexes_are_neighbors}
\title{Check whether two H3 indexes are neighbors (share an edge)}
\usage{
h3_indexes_are_neighbors(origin_h3_index, destination_h3_index)
}
\arguments{
\item{origin_h3_index}{H3 index (origin) [vector of indexes]}

\item{destination_h3_index}{H3 index (destination) [single index]}
}
\value{
boolean vector
}
\description{
js pendant: \code{h3.h3IndexesAreNeighbors(origin, destination) => boolean}
}
\examples{
coords <- c(37.3615593, -122.0553238)
h3_index <- geo_to_h3(coords)
neighbors <- k_ring(h3_index, 2)

h3_indexes_are_neighbors(neighbors, h3_index)
}
