% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_core.R
\name{h3_is_valid}
\alias{h3_is_valid}
\title{Check whether a given string represents a valid H3 index}
\usage{
h3_is_valid(h3_index)
}
\arguments{
\item{h3_index}{character vector representing H3 indexes}
}
\value{
boolean vector
}
\description{
Check whether a given string represents a valid H3 index
}
