% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_algorithms.R
\name{h3_set_to_multi_polygon}
\alias{h3_set_to_multi_polygon}
\title{Get the outlines of a set of hexagons}
\usage{
h3_set_to_multi_polygon(h3_index, format_as_geojson = TRUE)
}
\arguments{
\item{h3_index}{h3 index}

\item{format_as_geojson}{provide GeoJSON output: [lng, lat] and closed loops?}
}
\description{
Get the outlines of a set of hexagons
}
\examples{
coords <- c(37.3615593, -122.0553238)

geo_to_h3(coords) \%>\%
  k_ring(3) \%>\%
  h3_set_to_multi_polygon() \%>\%
  geo_boundary_to_sf() \%>\%
  plot()
}
