% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_h3-to-geo-boundary.R
\name{h3_to_geo_boundary}
\alias{h3_to_geo_boundary}
\title{Get the vertices of a given hexagon (or pentagon), as an array of [lat, lng] points}
\usage{
h3_to_geo_boundary(h3_index)
}
\arguments{
\item{h3_index}{h3 index}
}
\description{
Get the vertices of a given hexagon (or pentagon), as an array of [lat, lng] points
}
