% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_h3-to-geo-boundary.R
\name{h3_to_geo_boundary}
\alias{h3_to_geo_boundary}
\title{Get the vertices of a given hexagon (or pentagon)}
\usage{
h3_to_geo_boundary(h3_index, format_as_geojson = TRUE)
}
\arguments{
\item{h3_index}{character vector representing H3 indexes}

\item{format_as_geojson}{provide GeoJSON output: [lng, lat] and closed loops?}
}
\value{
matrix with [lat, lng] or [lng, lat] points if \code{format_as_geojson = TRUE};
Furthermore, GeoJSON format returns closed loops; list of matrices if multiple indexes
are supplied
}
\description{
Get the vertices of a given hexagon (or pentagon) either as an array of [lat, lng] points
or as GeoJSON output: [lng, lat] points and closed loops.
}
\seealso{
\link{geo_boundary_to_sf} and \link{geo_to_sf}
}
