% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_info.R
\name{num_hexagons}
\alias{num_hexagons}
\title{Get the total count of hexagons in the world at a given resolution}
\usage{
num_hexagons(res = 7)
}
\arguments{
\item{res}{resolution; numeric vector with values between 0 and 15}
}
\value{
numeric vector
}
\description{
Get the total count of hexagons in the world at a given resolution
}
\note{
For resolutions above 8 the exact count cannot be represented in a JavaScript 32-bit number.
  So use with caution when applying further operations to the output.
}
\examples{
num_hexagons(2:7)

}
