% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_polyfill.R
\name{polyfill}
\alias{polyfill}
\alias{polyfill.matrix}
\alias{polyfill.data.frame}
\alias{polyfill.list}
\alias{polyfill.sf}
\title{Get all hexagons with centers contained in a given polygon}
\usage{
polyfill(coords, res = 7, is_geojson = TRUE)

\method{polyfill}{matrix}(coords, res = 7, is_geojson = TRUE)

\method{polyfill}{data.frame}(coords, res = 7, is_geojson = TRUE)

\method{polyfill}{list}(coords, res = 7, is_geojson = TRUE)

\method{polyfill}{sf}(coords, res = 7, is_geojson = TRUE)
}
\arguments{
\item{coords}{numeric vector, matrix or data.frame with [lat, lng]
or [lng, lat] pairs if \code{is_geojson = TRUE};
object of class \code{sf} (geom type \code{POLYGON} or \code{MULTIPOLYGON})}

\item{res}{resolution; number between 0 and 15}

\item{is_geojson}{boolean; expect [lng, lat] pairs instead of [lat, lng]?}
}
\value{
h3 indexes; character vector or list of character vectors if multiple polygons
are supplied
}
\description{
Get all hexagons with centers contained in a given polygon
}
\examples{
library(sf)

list(greater_manchester) \%>\%
  st_polygon() \%>\%
  st_sfc(crs = 4326) \%>\%
  plot()

polyfill(greater_manchester, 6) \%>\%
  h3_to_geo_boundary() \%>\%
  geo_boundary_to_sf() \%>\%
  plot(add = TRUE)

nc <- system.file("shape/nc.shp", package = "sf") \%>\%
  st_read(quiet = TRUE)

st_geometry(nc) \%>\%
  plot(border = "red")

polyfill(nc, 5) \%>\% unlist() \%>\%
  h3_to_geo_boundary() \%>\%
  geo_boundary_to_sf() \%>\%
  plot(add = TRUE)
}
